Refactor: sortAZ.test.js :arrow_righ: sort.test.js
